Add assembly files + time differences between gcc and llvm
Downloaded and built TSVC benchmark: https://github.com/UoB-HPC/TSVC_2
Ran vectorized tests with relaxed floating point flags and compared the runtime between llvm and gcc. Output is stored in txt files and xlsx.

For the test case where llvm took the longest with respec to gcc, we go the assembly to compare. It looks like gcc was able to vectorized the code (used packing instructions) while llvm was not.